refactor: modal 공통 컴포넌트 분리 (common // challenge)
