+++ src/features/challenge/components/modal/camera-modal/index.ts
