+++ src/features/challenge/config/challenge-modal-provider/index.ts
@@ -1,0 +1,1 @@
+export * from './challenge-modal-provider'
