+++ src/features/common/components/index.ts
@@ -1,2 +1,1 @@
 export * from './global-wrapper'
-export * from './polling-watcher'
