+++ src/shared/components/modal/index.ts
@@ -1,5 +1,2 @@
-export * from './camera-modal'
 export * from './confirm-modal'
-export * from './image-zoom-modal'
 export * from './info-modal'
-export * from './verification-guide-modal'
